requirements.txt
django>=6.0,<7.0        # https://github.com/django/django

#Trivia site libraries
django-bulma>=0.8,<0.9
environs==9.5.0
whitenoise==6.5.0